+++ src/torchfem/__init__.py
@@ -5,6 +5,6 @@ warnings.filterwarnings("ignore", "Sparse CSR tensor support is in beta")
 from .assembly import Assembly, ReferencePoint, ReferencePointHeat  # noqa
 from .laminate import Laminate  # noqa
 from .planar import Planar, PlanarHeat  # noqa
-from .shell import Shell  # noqa
+from .shell import Shell, ShellHeat  # noqa
 from .solid import Solid, SolidHeat  # noqa
 from .truss import Truss  # noqa